requirements.txt
africastalking==1.2.5
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contextlib2==21.6.0
coverage==7.3.1
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
phonenumbers==8.13.20
psycopg2-binary==2.9.7
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
schema==0.7.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
wrapt==1.15.0
